Add files via upload
"Assigment_stat": this is a statistical model project  that involves the development and implementation of model and statistical tools to fulfill specific tasks and requirements outlined by the professor. The main objectives include data analysis, model construction, evaluation, and interpretation of the results.
"Paper_replication" : macroeconomics project that involves replicating the paper by Blanchard and Perotti, which explores the dynamic effects of fiscal policy on output. The goal is to understand the methodologies used, reproduce their results, and analyze the implications of their findings.